requirements.txt
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
python-decouple
pymongo==4.4.1